Add files via upload
Welcome to the project of Power BI! This Project Helped me to apply all that I've learned in this course.
Here's the problem statement:
Your client is a meal delivery company that operates in multiple cities. They have various fulfilment centres in these cities for dispatching meal orders to their customers. The client wants to help these centres with demand forecasting for upcoming weeks so that these centres will plan the stock of raw materials accordingly.
The replenishment of most raw materials is done on a weekly basis and since the raw material is perishable, procurement planning is of utmost importance. Secondly, staffing of the centres is also one area wherein accurate demand forecasts are helpful. We have the below information with us in the form of 3 different datasets:
•	Historical data of demand for a product-canter combination
•	Product (Meal) features such as category, sub-category, current price, and discount
•	Information for fulfilment centres like centre area, city information, etc.
Here is what we are expecting from you:
You need to come up with a story in Power BI that talks about the level of demand in each centre. This analysis needs to be granular enough to include product information as well. The client wants an end-to-end report to understand which fulfilment areas are doing well and which are not. You can also talk about centre-meal combinations to add nuance to your final submission.